feat: elasticsearch dirt user queue + other things
Former-commit-id: 7f24d95f289dc754c0a0665c00d218ffc957b303

+++ api/lib/flirtual/user/profile/custom_weights.ex
@@ -21,16 +21,16 @@ defmodule Flirtual.User.Profile.CustomWeights do
   schema "user_profile_custom_weights" do
     belongs_to :profile, Profile
 
-    field :country, :integer, default: 1
-    field :monopoly, :integer, default: 1
-    field :games, :integer, default: 1
-    field :default_interests, :integer, default: 1
-    field :custom_interests, :integer, default: 1
-    field :personality, :integer, default: 1
-    field :serious, :integer, default: 1
-    field :domsub, :integer, default: 1
-    field :kinks, :integer, default: 1
-    field :likes, :integer, default: 1
+    field :country, :float, default: 1.0
+    field :monopoly, :float, default: 1.0
+    field :games, :float, default: 1.0
+    field :default_interests, :float, default: 1.0
+    field :custom_interests, :float, default: 1.0
+    field :personality, :float, default: 1.0
+    field :serious, :float, default: 1.0
+    field :domsub, :float, default: 1.0
+    field :kinks, :float, default: 1.0
+    field :likes, :float, default: 1.0
   end
 
   def update_changeset(%Profile.CustomWeights{} = custom_weights, attrs) do
